+++ app/models/project_user.rb
@@ -6,7 +6,6 @@ class ProjectUser < ApplicationRecord
   include Forkable
 
   # Relationships
-  belongs_to :creator, class_name: "User", foreign_key: "creator_id", optional: true # TODO: Remove this line in v70+.
   belongs_to :project
-  belongs_to :user, optional: true # TODO: In v70+ should no longer be optional.
+  belongs_to :user
 end